Adapt encoder components to BoCA changes (declare supported input sample formats; remove obsolete checks and conversions).

+++ components/encoder/faac/faac.h
@@ -33,7 +33,7 @@ namespace BoCA
 			Int			 totalSamples;
 
 			Buffer<unsigned char>	 outBuffer;
-			Buffer<int32_t>		 samplesBuffer;
+			Buffer<int16_t>		 samplesBuffer;
 
 			Int			 EncodeFrames(Bool);
